스프링 카페 1-3 단계 구현
JdbcTemplateUserRepository 사용해서 h2 에 user 데이터 저장, 조회 구현

+++ src/main/java/com/kakao/cafe/web/controller/ArticleController.java
@@ -39,7 +39,7 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{index}")
-    public String showArticle(Model model, @PathVariable int index) {
+    public String showArticle(Model model, @PathVariable Long index) {
         logger.info("GET /articles/{}: response article detail page", index);
 
         // article 조회
